Update patch set 1
Patch Set 1:

I think we need have the CSIT test up aswell and try to get them both pushed at the same time as to avoid breakage in the test. We should also be updating the version of datarouter to 2.0.0 as users will have to update there code to get this working and it is no longer backward compatible. This is according to the standard of ONAP trying to use https://semver.org/

Patch-set: 1