Update patch set 5
Patch Set 5:

Eric Debeau confirms that the JSON is fine but does not render perfectly. For this requirement a link to your JSON is sufficient as mentioned in Doc-a-thon 4/17 by Eric. Code Patch 5 changes to just a link to our JSON and no rendering in the document.

Patch-set: 5
Label: Code-Review=0